swtpm: Cast constant to uint64_t before shifting (CID 375870)
Cast the '1' to uint64_t as suggested by Coverity (CID 375870).
Since 'j' is always less than '32', the previous code was correct as
well.

+++ src/swtpm_setup/swtpm.c
@@ -584,7 +584,7 @@ static int swtpm_tpm2_set_active_pcr_banks(struct swtpm *self, gchar **pcr_banks
             }
             pcrselects_len += len;
             count++;
-            activated_mask |= (1 << j);
+            activated_mask |= ((uint64_t)1 << j);
         }
     }
